pyproject.toml
[project]
name = "ftp-transfers"
version = "0.1.0"
description = "process images as they get transfered over sftp"
authors = [
    {name = "shadowlerone"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "watchdog (>=6.0.0,<7.0.0)",
    "pyexiftool (>=0.5.6,<0.6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "python-xmp-toolkit (>=2.0.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
